==> src/preprocess/01_cohort_selection.ipynb <==
import os
import sys

src_path = os.path.abspath('../..')
print(src_path)
sys.path.append(src_path)

from src.utils import create_directory, dump_pickle, raw_data_path, processed_data_path, set_seed

set_seed(seed=42)

import pandas as pd

mimic_iv_path = os.path.join(raw_data_path, "physionet.org/files/mimiciv/2.2")
mimic_iv_note_path = os.path.join(raw_data_path, "physionet.org/files/mimic-iv-note/2.2")

!ls {mimic_iv_path}

!ls {mimic_iv_note_path}

patients = pd.read_csv(os.path.join(mimic_iv_path, "hosp/patients.csv.gz"))
print(patients.shape)
patients.head()

admissions = pd.read_csv(os.path.join(mimic_iv_path, "hosp/admissions.csv.gz"))
print(admissions.shape)
admissions.head()

icustays = pd.read_csv(os.path.join(mimic_iv_path, "icu/icustays.csv.gz"))
print(icustays.shape)
icustays.head()

discharge = pd.read_csv(os.path.join(mimic_iv_note_path, "note/discharge.csv.gz"))
print(discharge.shape)
discharge.head()

print(patients.subject_id.nunique())
print(admissions.subject_id.nunique())
print(icustays.subject_id.nunique())
print(discharge.subject_id.nunique())

print(admissions.hadm_id.nunique())
print(icustays.hadm_id.nunique())
print(discharge.hadm_id.nunique())

print(icustays.stay_id.nunique())

admissions_hadm_ids = set(admissions.hadm_id.tolist())
icustays_hadm_ids = set(icustays.hadm_id.tolist())
discharge_hadm_ids = set(discharge.hadm_id.tolist())

from matplotlib_venn import venn3
import matplotlib.pyplot as plt


plt.figure(figsize=(8, 8))
venn3([admissions_hadm_ids, icustays_hadm_ids, discharge_hadm_ids], ('Hospital', 'ICU', 'Discharge'))
plt.show()

starting cohort

print(icustays.subject_id.nunique())
print(icustays.hadm_id.nunique())
print(icustays.stay_id.nunique())

remove no hospital admission

icustays = icustays[icustays.hadm_id.isin(admissions_hadm_ids)]

print(icustays.subject_id.nunique())
print(icustays.hadm_id.nunique())
print(icustays.stay_id.nunique())

remove no discharge note

icustays = icustays[icustays.hadm_id.isin(discharge_hadm_ids)]

print(icustays.subject_id.nunique())
print(icustays.hadm_id.nunique())
print(icustays.stay_id.nunique())

remove >2 ICU stays per admission

to_keep = icustays.groupby("hadm_id").stay_id.nunique().reset_index()
to_keep = to_keep[to_keep.stay_id == 1][["hadm_id"]]
icustays = icustays.merge(to_keep, how="inner", on="hadm_id")

print(icustays.subject_id.nunique())
print(icustays.hadm_id.nunique())
print(icustays.stay_id.nunique())

remove < 18 years old patients

icustays.intime = pd.to_datetime(icustays.intime)
icustays = icustays.merge(patients[["subject_id", "anchor_age", "anchor_year"]], on="subject_id", how="inner")
icustays["age"] = icustays.intime.dt.year - icustays.anchor_year + icustays.anchor_age
icustays.head()

icustays.age.min()

remove negative los

icustays = icustays.merge(admissions[["hadm_id", "admittime", "dischtime"]], on="hadm_id", how="inner")
icustays = icustays.rename(columns={
    "admittime": "hadm_intime",
    "dischtime": "hadm_outtime",
    "intime": "stay_intime",
    "outtime": "stay_outtime",
})
icustays.head()

icustays['hadm_intime'] = pd.to_datetime(icustays['hadm_intime'])
icustays['hadm_outtime'] = pd.to_datetime(icustays['hadm_outtime'])
icustays['stay_intime'] = pd.to_datetime(icustays['stay_intime'])
icustays['stay_outtime'] = pd.to_datetime(icustays['stay_outtime'])

icustays['stay_los'] = (icustays['stay_outtime'] - icustays['stay_intime']).dt.total_seconds() / 3600
icustays['hadm_los'] = (icustays['hadm_outtime'] - icustays['hadm_intime']).dt.total_seconds() / 3600

icustays.stay_los.min()

icustays.hadm_los.min()

icustays = icustays[icustays.stay_los >= 0]
icustays = icustays[icustays.hadm_los >= 0]

print(icustays.subject_id.nunique())
print(icustays.hadm_id.nunique())
print(icustays.stay_id.nunique())

statistics

icustays.groupby("subject_id").hadm_id.nunique().value_counts()

icustays.groupby("subject_id").hadm_id.nunique().describe()

icustays.groupby("hadm_id").stay_id.nunique().value_counts()

icustays.groupby("hadm_id").stay_id.nunique().describe()

icustays.stay_los.describe()

icustays.hadm_los.describe()

save

icustays.head()

icustays = icustays[["subject_id", "hadm_id", "stay_id", "hadm_intime", "hadm_outtime", "hadm_los", "stay_intime", "stay_outtime", "stay_los"]]
icustays.head()

output_path = os.path.join(processed_data_path, "mimic4")
create_directory(output_path)

icustays.to_csv(os.path.join(output_path, 'cohort.csv'), index=False)
