==> src/preprocess/02_event_hosp_microbiologyevents.ipynb <==
import os
import sys

src_path = os.path.abspath("../..")
print(src_path)
sys.path.append(src_path)

from src.utils import create_directory, raw_data_path, processed_data_path, set_seed

set_seed(seed=42)

import pandas as pd

mimic_iv_path = os.path.join(raw_data_path, "physionet.org/files/mimiciv/2.2")
output_path = os.path.join(processed_data_path, "mimic4")

cohort = pd.read_csv(os.path.join(output_path, "cohort.csv"))
print(cohort.shape)
cohort.head()

cohort["hadm_intime"] = pd.to_datetime(cohort["hadm_intime"])
cohort["hadm_outtime"] = pd.to_datetime(cohort["hadm_outtime"])
cohort["stay_intime"] = pd.to_datetime(cohort["stay_intime"])
cohort["stay_outtime"] = pd.to_datetime(cohort["stay_outtime"])

hadm_ids = set(cohort.hadm_id.unique().tolist())
len(hadm_ids)

helper

from concurrent.futures import ThreadPoolExecutor
from tqdm import tqdm
from pandarallel import pandarallel

pandarallel.initialize(progress_bar=True)

def save_group(group_df, hadm_id, event_type):
    file_path = f"{output_path}/event_{event_type}/event_{int(hadm_id)}.csv"
    group_df.to_csv(file_path, index=False)
    return True

dict

d_items = pd.read_csv(os.path.join(mimic_iv_path, "icu/d_items.csv.gz"))
print(d_items.shape)
d_items.head()

d_labitems = pd.read_csv(os.path.join(mimic_iv_path, "hosp/d_labitems.csv.gz"))
print(d_labitems.shape)
d_labitems.head()

## microbiologyevents

event_type = "microbiologyevents"

!rm -r {output_path}/event_{event_type}

create_directory(f"{output_path}/event_{event_type}")

df_raw = pd.read_csv(os.path.join(mimic_iv_path, f"hosp/{event_type}.csv.gz"))
print(df_raw.shape)
df_raw.head()

df = df_raw.merge(cohort[["hadm_id", "hadm_intime"]], on=["hadm_id"], how="inner")
print(df.shape)
df.head()

df["charttime"] = pd.to_datetime(df["charttime"])
df["storetime"] = pd.to_datetime(df["storetime"])
df["timestamp"] = (df.charttime - df.hadm_intime).dt.total_seconds() / 3600
df["timestamp_avail"] = (df.storetime - df.hadm_intime).dt.total_seconds() / 3600
df.head()

df = df.sort_values(["subject_id", "hadm_id", "timestamp", "spec_itemid"], ascending=True)
df.head()

df.isna().sum()

df = df.dropna(subset=["timestamp_avail"])
print(df.shape)
df.head()

df.interpretation.unique()

df.interpretation = df.interpretation.replace('S', 'sensitive')
df.interpretation = df.interpretation.replace('R', 'resistant')
df.interpretation = df.interpretation.replace('I', 'intermediate')
df.head()

def generate_event_value(x):
    s = f"{x.test_name} on {x.spec_type_desc}"
    if not pd.isna(x.org_name):
        s += f", organism grew: {x.org_name}"
    if not pd.isna(x.ab_name):
        s += f", antibiotic tested: {x.ab_name}"
    if not pd.isna(x.interpretation):
        s += f", antibiotic sensitivity: {x.interpretation}"
    if not pd.isna(x.comments):
        s += f", comments: {x.comments}"
    return s

meta_cols = ["test_name", "spec_type_desc", "org_name", "ab_name", "interpretation", "comments"]
for c in meta_cols:
    df["meta_" + c] = df[c]
meta_cols = ["meta_" + c for c in meta_cols]

df["event_type"] = event_type
df["event_value"] = df.parallel_apply(generate_event_value, axis=1)

df[df.hadm_id == 29079034]

df.groupby("hadm_id").timestamp.count().describe()

groups = df.groupby("hadm_id")
    
with ThreadPoolExecutor(max_workers=4) as executor:
    for hadm_id, group_df in tqdm(groups, total=groups.ngroups):
        future = executor.submit(
            save_group, 
            group_df[["hadm_id", "event_type", "timestamp", "event_value", "timestamp_avail"] + meta_cols], 
            hadm_id, 
            event_type
        )

!ls -1 {output_path}/event_{event_type} | wc -l
